+++ mods/CreateDragonPlus/src/main/java/plus/dragons/createdragonsplus/client/ponder/CDPPonderScenes.java
@@ -21,11 +21,19 @@ package plus.dragons.createdragonsplus.client.ponder;
 import com.simibubi.create.AllBlocks;
 import com.tterrag.registrate.util.entry.ItemProviderEntry;
 import com.tterrag.registrate.util.entry.RegistryEntry;
+import java.util.List;
+import net.createmod.catnip.registry.RegisteredObjectsHelper;
 import net.createmod.ponder.api.registration.PonderSceneRegistrationHelper;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.ItemLike;
 import plus.dragons.createdragonsplus.client.ponder.scenes.CDPFanScenes;
+import plus.dragons.createdragonsplus.client.ponder.scenes.CDPFluidScenes;
 import plus.dragons.createdragonsplus.client.ponder.scenes.SandingScenes;
+import plus.dragons.createdragonsplus.common.fluids.dye.DyeVariantRegistry;
 import plus.dragons.createdragonsplus.common.kinetics.fan.sanding.SandingCatalysts;
+import plus.dragons.createdragonsplus.common.registry.CDPCauldrons;
+import plus.dragons.createdragonsplus.common.registry.CDPFluids;
 
 public class CDPPonderScenes {
     public static void register(PonderSceneRegistrationHelper<ResourceLocation> helper) {
@@ -39,5 +47,19 @@ public class CDPPonderScenes {
             registration.forComponents(AllBlocks.ENCASED_FAN)
                     .addStoryBoard("bulk_sanding", SandingScenes::bulkSanding);
         }
+
+        var itemRegistration = helper.<ItemLike>withKeyFunction(RegisteredObjectsHelper::getKeyOrThrow);
+        List<ItemLike> dyeBuckets = DyeVariantRegistry.all().stream()
+                .filter(variant -> variant.isAvailable())
+                .flatMap(variant -> CDPFluids.DYES_BY_VARIANT.get(variant.id()).getBucket().stream())
+                .map(bucket -> (ItemLike) bucket)
+                .toList();
+        itemRegistration.forComponents(dyeBuckets)
+                .addStoryBoard("dye_fluids", CDPFluidScenes::dyeFluids);
+        itemRegistration.forComponents(
+                CDPFluids.DRAGON_BREATH.getBucket().get(),
+                CDPCauldrons.DRAGON_BREATH_CAULDRON.get(),
+                Items.DRAGON_BREATH)
+                .addStoryBoard("dragon_breath_fluid", CDPFluidScenes::dragonBreathFluid);
     }
 }