+++ mods/CreateDragonPlus/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Create: Dragons Plus 1.11.2
 
+### Update
+* Added Ponder for Bulk Enchanting, Fluid Dye, Dragon's Breath.
+
 ### Fix
 * Fixed Bulk Coloring air current block interaction resetting the blockstate of colored blocks.